requirements.txt
streamlit
pandas
python-dotenv
openpyxl
google-generativeai
tabulate
